chore(template): refactor pdp component depth
